==> src/01_data_imports_and_manipulation.ipynb <==
# Importing the Raw Datasets

import pandas as pd

wdi_df = pd.read_csv("./data/WDIData.csv")

# Output the first 100 records as a sample set
wdi_df.head(100).to_csv("./data/WDIData_sample.csv")

wdi_df.head()


climate_df = pd.read_csv("./data/climate_watch_data.csv")

climate_df.head()


un_sds_df = pd.read_excel("./data/un_sds_poverty_data.xlsx")

un_sds_df.head()


# Matching Country Names

Some countries do not share a common name between the datasets. The following code was used to generate lists of unique country names not included in the other datasets.

Since the GHG data contained the smallest number of country names it was chosen as the “source of truth” for country labels. 


climate_countries = climate_df["Country"].unique()
wdi_countries = wdi_df["Country Name"].unique()
un_sds_countries = sorted(un_sds_df["GeoAreaName"].unique())

in_climate_not_wdi = list(set(climate_countries).difference(set(wdi_countries)))
in_climate_not_un = list(set(climate_countries).difference(set(un_sds_countries)))


Mapping GHG Country names to other datasets was not a process that led itself to automation. In the case of countries like `Kyrgyz Republic/Kyrgyzstan` it might've been possible, but difficult to use partial matching to connect the names together. With a total list of 34 country names, this was a task that we undertook manually with discretion.


| Greenhouse Gas Emissions         | World Development Indicators (GDP Data) | Global Poverty Data                                      |
| -------------------------------- | --------------------------------------- | -------------------------------------------------------- |
| Bahamas                          | **Bahamas, The**                        | Bahamas                                                  |
| Bolivia                          | Bolivia                                 | **Bolivia (Plurinational State of)**                     |
| Brunei                           | **Brunei Darussalam**                   | **Brunei Darussalam**                                    |
| Cape Verde                       | **Cabo Verde**                          | **Cabo Verde**                                           |
| Cook Islands                     | _N/A_                                   | Cook Islands                                             |
| Côte d'Ivoire                    | **Cote d'Ivoire**                       | Côte d'Ivoire                                            |
| Czech Republic                   | Czech Republic                          | **Czechia**                                              |
| Democratic Republic of the Congo | **Congo, Dem. Rep.**                    | Democratic Republic of the Congo                         |
| Egypt                            | **Egypt, Arab Rep.**                    | Egypt                                                    |
| European Union (27)              | **European Union**                      | _N/A_                                                    |
| Gambia                           | **Gambia, The**                         | Gambia                                                   |
| Iran                             | **Iran, Islamic Rep.**                  | **Iran (Islamic Republic of)**                           |
| Kyrgyzstan                       | **Kyrgyz Republic**                     | Kyrgyzstan                                               |
| Laos                             | **Lao PDR**                             | **Lao People's Democratic Republic**                     |
| Macedonia                        | **North Macedonia**                     | **North Macedonia**                                      |
| Micronesia                       | **Micronesia, Fed. Sts.**               | Micronesia                                               |
| Moldova                          | Moldova                                 | **Republic of Moldova**                                  |
| Niue                             | _N/A_                                   | Niue                                                     |
| North Korea                      | **Korea, Dem. People's Rep.**           | **Democratic People's Republic of Korea**                |
| Republic of Congo                | **Congo, Rep.**                         | **Congo**                                                |
| Russia                           | **Russian Federation**                  | **Russian Federation**                                   |
| Saint Kitts and Nevis            | **St. Kitts and Nevis**                 | Saint Kitts and Nevis                                    |
| Saint Lucia                      | **St. Lucia**                           | Saint Lucia                                              |
| Saint Vincent and the Grenadines | **St. Vincent and the Grenadines**      | Saint Vincent and the Grenadines                         |
| Slovakia                         | **Slovak Republic**                     | Slovakia                                                 |
| South Korea                      | **Korea, Rep.**                         | **Republic of Korea**                                    |
| Syria                            | **Syrian Arab Republic**                | **Syrian Arab Republic**                                 |
| Tanzania                         | Tanzania                                | **United Republic of Tanzania**                          |
| Turkey                           | **Turkiye**                             | **Türkiye**                                              |
| United Kingdom                   | United Kingdom                          | **United Kingdom of Great Britain and Northern Ireland** |
| United States                    | United States                           | **United States of America**                             |
| Venezuela                        | **Venezuela, RB**                       | **Venezuela (Bolivarian Republic of)**                   |
| Vietnam                          | Vietnam                                 | **Viet Nam**                                             |
| Yemen                            | **Yemen, Rep**                          | Yemen                                                    |


climate_rename_list = [
    "Bahamas",
    "Bolivia",
    "Brunei",
    "Cape Verde",
    "Cook Islands",
    "Côte d'Ivoire",
    "Czech Republic",
    "Democratic Republic of the Congo",
    "Egypt",
    "European Union (27)",
    "Gambia",
    "Iran",
    "Kyrgyzstan",
    "Laos",
    "Macedonia",
    "Micronesia",
    "Moldova",
    "Niue",
    "North Korea",
    "Republic of Congo",
    "Russia",
    "Saint Kitts and Nevis",
    "Saint Lucia",
    "Saint Vincent and the Grenadines",
    "Slovakia",
    "South Korea",
    "Syria",
    "Tanzania",
    "Turkey",
    "United Kingdom",
    "United States",
    "Venezuela",
    "Vietnam",
    "Yemen",
]

wdi_rename_list = [
    "Bahamas, The",
    "Bolivia",
    "Brunei Darussalam",
    "Cabo Verde",
    "",
    "Cote d'Ivoire",
    "Czech Republic",
    "Congo, Dem. Rep.",
    "Egypt, Arab Rep.",
    "European Union",
    "Gambia, The",
    "Iran, Islamic Rep.",
    "Kyrgyz Republic",
    "Lao PDR",
    "North Macedonia",
    "Micronesia, Fed. Sts.",
    "Moldova",
    "",
    "Korea, Dem. People's Rep.",
    "Congo, Rep.",
    "Russian Federation",
    "St. Kitts and Nevis",
    "St. Lucia",
    "St. Vincent and the Grenadines",
    "Slovak Republic",
    "Korea, Rep.",
    "Syrian Arab Republic",
    "Tanzania",
    "Turkiye",
    "United Kingdom",
    "United States",
    "Venezuela, RB",
    "Vietnam",
    "Yemen, Rep.",
]

un_sds_rename_list = [
    "Bahamas",
    "Bolivia (Plurinational State of)",
    "Brunei Darussalam",
    "Cabo Verde",
    "Cook Islands",
    "Côte d'Ivoire",
    "Czechia",
    "Democratic Republic of the Congo",
    "Egypt",
    "",
    "Gambia",
    "Iran (Islamic Republic of)",
    "Kyrgyzstan",
    "Lao People's Democratic Republic",
    "North Macedonia",
    "Micronesia",
    "Republic of Moldova",
    "Niue",
    "Democratic People's Republic of Korea",
    "Congo",
    "Russian Federation",
    "Saint Kitts and Nevis",
    "Saint Lucia",
    "Saint Vincent and the Grenadines",
    "Slovakia",
    "Republic of Korea",
    "Syrian Arab Republic",
    "United Republic of Tanzania",
    "Türkiye",
    "United Kingdom of Great Britain and Northern Ireland",
    "United States of America",
    "Venezuela (Bolivarian Republic of)",
    "Viet Nam",
    "Yemen",
]


# Generating the Clean Output

Generating the list of years that we are focusing on for our analysis, 1990-2019.

years = list(map(str, range(1990, 2020)))


wdi_df_clean = wdi_df[
    [
        "Country Name",
        "Country Code",
        "Indicator Name",
        "Indicator Code",
        *years,
    ]
]

wdi_rename_dict = dict(zip(wdi_rename_list, climate_rename_list))
wdi_df_clean = wdi_df_clean.replace(wdi_rename_dict).dropna(subset=years, how="all")
wdi_df_clean = wdi_df_clean[wdi_df_clean["Country Name"].isin(climate_countries)]

# Uncomment to overwrite clean output
# wdi_df_clean.to_csv("./data/wdi_clean.csv", index=False)

# Output the first 100 records as a sample set
wdi_df_clean.head(100).to_csv("./data/wdi_clean_sample.csv", index=False)

wdi_df_clean.head()


The UN SDS Poverty data required some extra manipulation here. We dropped irrelevant columns and pivoted the data from a long to wide format to match the other two datasets.


# fmt:off
un_sds_clean = un_sds_df[
    un_sds_df["TimePeriod"].ge(1990) & un_sds_df["TimePeriod"].le(2019) & ~un_sds_df["Quantile"].eq("Q1")
].drop(["Goal", "Target", "Indicator", "GeoAreaCode", "Time_Detail", "TimeCoverage", "UpperBound", "LowerBound", "BasePeriod", "Source", "GeoInfoUrl", "FootNote", "Hazard type", "Quantile", "Nature", "Observation Status", "Reporting Type"], axis="columns" )
# fmt:on

un_sds_rename_dict = dict(zip(un_sds_rename_list, climate_rename_list))
un_sds_clean = un_sds_clean.replace(un_sds_rename_dict)
un_sds_clean = un_sds_clean[un_sds_clean["GeoAreaName"].isin(climate_countries)]


un_sds_clean_pivot = un_sds_clean.pivot(
    index=[
        "SeriesCode",
        "SeriesDescription",
        "GeoAreaName",
        "Age",
        "Location",
        "Sex",
        "Units",
    ],
    columns="TimePeriod",
    values="Value",
)

# Uncomment to overwrite clean output
# un_sds_clean_pivot.to_csv("./data/un_sds_clean.csv")

un_sds_clean_pivot.head()


%reload_ext watermark

%watermark -iv -v -m
