==> src/02_combining_datasets.ipynb <==
# Combining Datasets


import pandas as pd
import numpy as np

wdi_df = pd.read_csv("./data/wdi_clean.csv")

wdi_df.head()


climate_df = pd.read_csv("./data/climate_watch_data.csv")

climate_df.head()


un_sds_df = pd.read_csv("./data/un_sds_clean.csv")

un_sds_df.head()


# Put Datasets into Common Column Configuration

| Column     | Description                                                                |
| ---------- | -------------------------------------------------------------------------- |
| country    | Country name                                                               |
| c_code     | Three letter shortcode for the country                                     |
| indicator  | Name for the value, for GHG data this describes the source of the emission |
| i_code     | Shorthand code for the indicator                                           |
| gas        | Type of emission gas (“CO2", “CH4”, etc.)                                  |
| 1990..2019 | Range of years for the data                                                |

Generating the list of years that we are focusing on for our analysis, 1990-2019.


years = list(map(str, range(1990, 2020)))


wdi_df_renamed = wdi_df.rename(
    {
        "Country Name": "country",
        "Country Code": "c_code",
        "Indicator Name": "indicator",
        "Indicator Code": "i_code",
    },
    axis="columns",
)

wdi_df_renamed.insert(4, "gas", "")

wdi_df_renamed.head()


codes = dict(zip(wdi_df_renamed["country"].unique(), wdi_df_renamed["c_code"].unique()))


def set_country_code(row):
    try:
        row["c_code"] = codes[row["country"]]
    except:
        pass
    return row


un_sds_df["indicator"] = (
    un_sds_df["SeriesDescription"]
    + un_sds_df["Units"].astype(str)
    + un_sds_df["Age"].replace({np.NaN: ""})
    + un_sds_df["Location"].replace({np.NaN: ""})
    + un_sds_df["Sex"].replace({np.NaN: ""})
)


un_sds_df_renamed = un_sds_df.rename(
    {"GeoAreaName": "country", "SeriesCode": "i_code"},
    axis="columns",
)

un_sds_df_renamed = un_sds_df_renamed.apply(set_country_code, axis="columns").reindex(
    [
        "country",
        "c_code",
        "indicator",
        "i_code",
        *years,
    ],
    axis="columns",
)


un_sds_df_renamed.insert(4, "gas", "")

un_sds_df_renamed.head()


climate_df_renamed = climate_df.drop(["Data source", "Unit"], axis="columns").rename(
    {
        "Country": "country",
        "Sector": "indicator",
        "Gas": "gas",
    },
    axis="columns",
)

climate_df_renamed = climate_df_renamed.apply(set_country_code, axis="columns").reindex(
    ["country", "c_code", "indicator", "gas", *years], axis="columns"
)

climate_df_renamed.insert(3, "i_code", "")

climate_df_renamed.head()


For the WDI data we are filtering this down to only GDP data which is relevant to our analysis.


wdi_df_gdp = wdi_df_renamed[wdi_df_renamed["indicator"].str.contains("gdp", case=False)]

combined_df = pd.concat(
    [climate_df_renamed, un_sds_df_renamed, wdi_df_gdp]
).sort_values(["country", "indicator", "gas"])

combined_df.head()


# len(combined_df["indicator"].unique())
# len(combined_df["country"].unique())
combined_df.columns


Our combined dataset contains `34,667` rows and is about `10.9 MB` as a csv.


# Uncomment to replace the complete_dataset
# combined_df.to_csv("./data/complete_dataset.csv", index=False)

# Output the first 100 records as a sample set
combined_df.head(100).to_csv("./data/complete_dataset_sample.csv", index=False)


%reload_ext watermark

%watermark -iv -v -m
