==> src/05_Top_10_Emitting_Countries.ipynb <==
# Import Modules
import pandas as pd
import numpy as np
import altair as alt
import warnings
import re


# Set the notebook to display all columns of a dataframe
pd.set_option("display.max_columns", None)

# Suppress warnings for clean cell outputs
warnings.filterwarnings("ignore")


# Load Data

# Load in cleaned & combined data
df = pd.read_csv("./data/complete_dataset_with_interpolation.csv")


# **Explore the cleaned dataset**

We're going to start by looking at the combined dataset. This cleaned dataset includes the `World Banks's Development Indicators`, `Climate Watch Greenhouse Gas Emissions Data`, and the `UN's Sustainable Development Poverty Data`. Recording observations on the structure of this data will guide further analysis.

# Look at values of GHG emissions that are not NaN
df[df["gas"].notna()].sample(5)


# Examine records for the poverty variable
df[
    df["indicator"]
    == "Proportion of population below international poverty line (%)PERCENTALLAGEALLAREABOTHSEX"
].head()


# Look at unique Greenhouse Gases
df["gas"].unique()


# Look at the types of sectors that emit GHG
df[df["gas"].notna()]["indicator"].unique()


### Observations 
* Data is broken down for various GHG types in the climate watch data. To maintain consistency - we will analyze relationships in our combined dataset using only `All GHG` climate watch data. 
* It looks like the GHG climate watch data is also broken down by sector of a country's economy. There may be some interesting relationships that exist within these sub groups of emissions data.

# Clean Country values of the Democratic Republic of the Congo
df["country"] = df["country"].apply(
    lambda x: "Democratic Republic of the Congo" if x == "Republic of Congo" else x
)

# Make a copy of the dataframe so the
# original one is still in the workspace and
# remains unaltered.
combined_df = df.copy()

# Drop the World and European Union records
to_drop = ["European Union (27)", "World"]
combined_df = combined_df[~combined_df["country"].isin(to_drop)]

# Filter data to grab only the variables we care about
# List of variables to grab from the dataset
# Dict to rename variable labels
vars = {
    "Proportion of population below international poverty line (%)PERCENTALLAGEALLAREABOTHSEX": "Percent_Poverty",
    "Total including LUCF, All GHG": "Total_GHG",
    "Total excluding LUCF, All GHG": "Total_GHG_exc",
    "GDP, PPP (current international $)": "GDP_Total",
    "Population, total": "Total_Population",
}


def get_long_form(to_melt_df, vars_of_interest, var_dict):
    """
    Converts a wide-format dataframe into a long form using
    the specified variables in vars_of_interest.

    Parameters:

         to_melt_df: dataframe in wide format.

         vars_of_interest: list of variables to include in the
         long -form dataframe. Should exist within the combined_df.

         var_dict: a dictionary to rename the variables in vars_of_interest

    Returns:

         dataframe: in long format

    """
    # Combine gas and indicator field
    to_melt_df["indicator"] = to_melt_df.apply(
        lambda row: row["indicator"] + ", " + row["gas"]
        if pd.notnull(row["gas"])
        else row["indicator"],
        axis=1,
    )

    # Filter the dataframe to grab vars of interest
    interest_df = to_melt_df[to_melt_df["indicator"].isin(vars_of_interest)]

    # Get the years in the dataframe
    pat = "[\d]{4}"
    year_cols = [col for col in interest_df.columns if re.search(pat, col)]

    # Melt the interest_df to long form
    out_dat = interest_df.melt(
        id_vars=["country", "c_code", "indicator"],
        value_vars=year_cols,
        var_name="Year",
        value_name="Value",
    )

    # Rename column labels
    out_dat = out_dat.replace(var_dict)

    # Groupby the indicator-year-country combos and get the mean
    # value to avoid duplicates in the index
    out_dat = (
        out_dat.groupby(["country", "c_code", "indicator", "Year"])["Value"]
        .mean()
        .reset_index()
    )

    # Pivot the dataframe to get the variables
    # as the column labels
    out_dat = out_dat.pivot(
        index=["country", "c_code", "Year"], columns=["indicator"], values=["Value"]
    ).reset_index()

    # Drop multi-levels in the columsn from the pivot
    out_dat.columns = out_dat.columns.droplevel(0)
    out_dat = out_dat.rename_axis(None, axis=1)

    # Rename empty column labels
    out_dat.columns = ["Country", "c_code", "Year"] + list(out_dat.columns[3:])

    # Convert year column to type int
    out_dat["Year"] = out_dat["Year"].astype(int)

    return out_dat

out_df = get_long_form(combined_df, vars.keys(), vars)

# View the result
out_df.head()


# Send long form of the data to CSV so it can be referenced in future notebooks

# out_df.to_csv("./data/complete_dataset_LongFormat.csv", index=False)


##  **Examine Missing Values for each variable and year**
Here I will call a custom groupby function (defined below) that gives a total count of non-null values for each variable in `out_df`. This will help give us an idea of where we have missing data in the dataframe. 

def summarize_na(dat):
    """
    Summarize function to be used in a groupby and
    determine the number of non-missing data in each
    column of our dataframe.

    Parameters:

        dataframe: a dataframe with columns to summarize missing values

    Returns:

        Series: series that contains the summary of missing values
    """

    # Use a dictionary comprehension to summarize
    # the number of non-missing values in each column
    out_dict = {col: len(dat[dat[col].notnull()]) for col in dat.columns}

    return pd.Series(out_dict, index=out_dict.keys())


# Look at counts of non-NaN protected values values by year
count_by_year = out_df.groupby("Year").apply(summarize_na)
count_by_year.tail(20)


### Observations of missing values:
* SDS poverty data is sparse for the first decade of the 21st century. 
* More poverty data is available from 2010 onward, but still only ~25% of all countries have this data. 
* We have good coverage for all years and countries total population, GHG data, and GDP. 

### **Quantify Emissions Reductions by Country**
I'm wondering about which countries have decreased their emissions most from their peak values. Below I group-by country and calculate the change in greenhouse gas emissions between the peak value and the 2019 value. Questions I hope to answer here include:

1. Which countries have reduced their emissions _**the most**_?
2. Of these countries that have decreased emissions from their peak values, how many of them are high GDP countries or low GDP countries?

def get_GHG_delta(dat):
    # Funtion that is to be used in the groupby
    # of countries to calculate the difference
    # in peak emissions and the current value

    # Get the max emissions data for the country
    record_peak_emissions = dat[dat["Total_GHG_exc"] == dat["Total_GHG_exc"].max()]

    # Grab the GHG value and the year of
    # peak emissions
    peak_emissions_val = record_peak_emissions["Total_GHG_exc"].iloc[0]
    peak_emissions_yr = int(record_peak_emissions["Year"].iloc[0])

    # Get the most recent emissions data
    emissions_recent = dat[dat["Year"] == dat["Year"].max()]

    # Grab the GHG value and the year of
    # most recent emissions
    emissions_recent_val = emissions_recent["Total_GHG_exc"].iloc[0]
    emission_recent_yr = int(emissions_recent["Year"].iloc[0])

    # Get the delta between peak and most recent
    # emissions amounts
    delta = peak_emissions_val - emissions_recent_val

    # output dictionary with data
    out_dict = {
        "Peak_Emissions_Yr": peak_emissions_yr,
        "Peak_Emissions_Val": peak_emissions_val,
        "Most_Recent_Yr": emission_recent_yr,
        "Recent_Emission_Val": emissions_recent_val,
        "Emissions_Delta": delta,
    }

    return pd.Series(out_dict, index=list(out_dict.keys()))


# Groupby Country and Apply the Custom Function
reductions_df = out_df.groupby("Country").apply(get_GHG_delta).reset_index()

# View Results
reductions_df.sample(5)


### **Visualize Emissions Reductions**

def get_reductions_chart():
    chart_df = reductions_df[reductions_df["Emissions_Delta"] > 0]

    chart_df.sort_values("Emissions_Delta", ascending=False, inplace=True)

    sort_order = [con for con in chart_df["Country"].unique()]

    chart_df = chart_df.iloc[:10]

    # Specs for left visual bar chart
    bars = (
        alt.Chart(chart_df)
        .mark_bar()
        .encode(
            y=alt.Y("Country:N", sort=sort_order, axis=alt.Axis(labelFontSize=12)),
            x=alt.X(
                "Emissions_Delta:Q",
                axis=alt.Axis(
                    title="Reductions in GHG Since Peak Emissions (MtCO2, excluding LUCF)",
                    titleFontSize=12,
                    labelFontSize=12,
                ),
            ),
        )
    )

    text = bars.mark_text(align="left", baseline="middle", dx=3).encode(
        y=alt.Y("Country:N", axis=alt.Axis(title=None), sort=sort_order),
        x=alt.X("Emissions_Delta:Q"),
        text=alt.Text("Emissions_Delta:Q", format=".0f"),
    )

    out = (
        alt.layer(bars, text)
        .properties(
            title={
                "text": "Reductions in Greenhouse Gasses since Peak Emissions by Country"
            },
        )
        .configure_title(fontSize=14, align="left")
        .configure_text(fontSize=12)
    )

    return out, sort_order


chart, order = get_reductions_chart()

chart


## **Identify Top 10 GHG emitting countries**
### First: prep the data for the visualization.

var = "Total_GHG"

# Filter for only data from 2009-2019
plot_df = out_df[out_df["Year"] >= 2010]

# Grab columns to summarize
plot_df = plot_df[["Country", "GDP_Total", "Total_GHG"]]

# Grab the mean over those years
plot_df = plot_df.groupby("Country").agg(np.mean).reset_index()

# edit the field specifying GDP to show $ amount in billions
plot_df["GDP_Total"] = plot_df["GDP_Total"] / 1000000000

# Grab top 15 bottom 15 countries
top_15 = plot_df.sort_values(var, ascending=False)
sort_top = [country for i, country in enumerate(top_15["Country"]) if i < 10]
bottom_15 = plot_df.sort_values(var, ascending=True).iloc[:14]
sort_bottom = [country for i, country in enumerate(bottom_15["Country"]) if i < 10]


# Melt the data so the values are in one column
viz_df = plot_df.melt(
    id_vars=["Country"],
    value_vars=["GDP_Total", "Total_GHG"],
    var_name="Parameter",
    value_name="Value",
)

viz_df = viz_df.replace(
    {"GDP_Total": "GDP PPP (billions $)", "Total_GHG": "Emissions (MtCO2)"}
)

# Redefine the to & bottom dataframes to get the
# countries we are interested in
top_15 = viz_df[viz_df["Country"].isin(sort_top)]
bottom_15 = viz_df[viz_df["Country"].isin(sort_bottom)]


### Second: Make the visual.

# Use altair to render the visualization
out_chart = (
    alt.Chart(top_15)
    .mark_bar()
    .encode(
        x="Value:Q",
        y=alt.Y("Parameter:N", axis=alt.Axis(title=None)),
        color="Parameter:N",
    )
    .facet(
        facet=alt.Facet(
            "Country:N",
            sort=sort_top,
            title=None,
            header=alt.Header(labelFontSize=14, labelFontWeight="bold"),
        ),
        columns=1,
        spacing=4,
    )
    .properties(
        title={
            "text": "Emissions and GDP Values for the top 10 Greenhouse Gas Emitting Countries",
            "subtitle": "Emissions and GDP values shown are the average of values between 2010 and 2019",
        }
    )
    .configure_title(fontSize=16, subtitleFontSize=12)
    .configure_legend(
        titleFontSize=14,
        labelFontSize=12,
        titlePadding=10,
        strokeColor="gray",
        padding=10,
    )
    .configure_header(
        labelAnchor="middle", labelOrient="right", labelAngle=0, labelPadding=20
    )
)

out_chart


%reload_ext watermark

%watermark -iv -v -m
